Create the classic Conway's Game of Life.
The user picks a starting shape, the size of the grid, and the number of turns. The 4 rules are: A living cell with 0 or 1 living neighbors dies, A living cell with 2 or 3 living neighbors lives, A living cell with 4 or more living neighbors dies, and A dead cell with exactly 3 living neighbors comes back to life. The game will play the number of turns by applying the 4 game rules to the current game board in order to generate the next gameboard. The results are displayed on the form and are repeated until all turns are played.